Refactor SyncActivity to extract preferences and server defaults
* Inject `beta_auto_download` preference retrieval to `SharedPrefManager`.
* Move `isLocalNetwork` and `getDefaultProtocol` logic from `ServerDialogExtensions` to `ServerConfigUtils`.
* Remove direct `PreferenceManager` access from `SyncActivity`.
