+++ app/src/main/java/org/ole/planet/myplanet/utils/ServerConfigUtils.kt
@@ -63,6 +63,25 @@ object ServerConfigUtils {
         return pinMap[url] ?: ""
     }
 
+    private fun isLocalNetwork(url: String): Boolean {
+        val host = url.split(":").firstOrNull()?.split("/")?.firstOrNull() ?: url
+        return host.startsWith("192.168.") ||
+                host.startsWith("10.") ||
+                host.matches(Regex("^172\\.(1[6-9]|2[0-9]|3[0-1])\\..*")) ||
+                host == "localhost" ||
+                host == "127.0.0.1" ||
+                host.endsWith(".local")
+    }
+
+    fun getDefaultProtocol(url: String): String {
+        return if (
+            url == BuildConfig.PLANET_XELA_URL ||
+            url == BuildConfig.PLANET_SANPABLO_URL ||
+            url == BuildConfig.PLANET_URIUR_URL ||
+            isLocalNetwork(url)
+        ) org.ole.planet.myplanet.utils.Constants.HTTP_PROTOCOL else org.ole.planet.myplanet.utils.Constants.HTTPS_PROTOCOL
+    }
+
     fun saveAlternativeUrl(
         url: String,
         password: String,
